// types/RegisteredDeviceResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Model } from "./DataValue";
import type { Model as DeviceModel } from "./Device";

export type RegisteredDeviceResponse = {
  device: DeviceModel;
  latest_data: Model | null;
};